requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.1.0
Babel==2.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django
django-cors-headers==4.7.0
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pillow==11.1.0
propcache==0.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
